rename av pakkenavn (#3462)
