+++ src/main/kotlin/no/nav/familie/ba/sak/integrasjoner/økonomi/internkonsistensavstemming/InternKonsistensavstemmingService.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.ba.sak.integrasjoner.økonomi.InternKonsistendsavstemming
+package no.nav.familie.ba.sak.integrasjoner.økonomi.internkonsistensavstemming
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
