+++ src/main/kotlin/no/nav/familie/ba/sak/integrasjoner/økonomi/internkonsistensavstemming/InternKonsistensavstemmingUtil.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.ba.sak.integrasjoner.økonomi.InternKonsistendsavstemming
+package no.nav.familie.ba.sak.integrasjoner.økonomi.internkonsistensavstemming
 
 import no.nav.familie.ba.sak.common.Utils
 import no.nav.familie.ba.sak.common.secureLogger
