+++ src/main/kotlin/no/nav/familie/ba/sak/task/InternKonsistensavstemmingTask.kt
@@ -1,7 +1,7 @@
 package no.nav.familie.ba.sak.task
 
 import com.fasterxml.jackson.module.kotlin.readValue
-import no.nav.familie.ba.sak.integrasjoner.økonomi.InternKonsistendsavstemming.InternKonsistensavstemmingService
+import no.nav.familie.ba.sak.integrasjoner.økonomi.internkonsistensavstemming.InternKonsistensavstemmingService
 import no.nav.familie.kontrakter.felles.objectMapper
 import no.nav.familie.log.mdc.MDCConstants
 import no.nav.familie.prosessering.AsyncTaskStep
